Completely redid the satisfiability checker.
Now, it solves exactly using integer constraints rather than using
linear systems. Not only is this sound, but is using recent developments
in ordering constraint SMT solving to do so.

We don't actually use an SMT solver, rather just the basic
graph-theoretical portion of the paper referenced in the description.